PR review fixes + profile pic default image file moved
